service/research/src/requirements.txt
psycopg2-binary==2.9.8
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pymongo==4.5.0
python-dotenv==1.0.0
